Make a new file for class
